BACKPORT: sched: Introduce latency-nice as a per-task attribute
Latency-nice indicates the latency requirements of a task with respect
to the other tasks in the system. The value of the attribute can be within
the range of [-20, 19] both inclusive to be in-line with the values just
like task nice values.

Just like task nice, -20 is the 'highest' priority and conveys this
task should get minimal latency, conversely 19 is the lowest priority
and conveys this task will get the least consideration and will thus
receive maximal latency.

[peterz: rebase, squash]

[xMikkkaa: Backported to v4.9. Manually introduced latency_prio in task_struct and sched_attr due to uapi layout differences, ensuring compatibility with legacy struct sched_entity.]

Signed-off-by: Parth Shah <parth@linux.ibm.com>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Tested-by: Shrikanth Hegde <sshegde@linux.vnet.ibm.com>
Signed-off-by: xMikkkaa <xMikkkaa@users.noreply.github.com>