BACKPORT: sched/fair: Add sched group latency support
Task can set its latency priority with sched_setattr(), which is then used
to set the latency offset of its sched_enity, but sched group entities
still have the default latency offset value.

Add a latency.nice field in cpu cgroup controller to set the latency
priority of the group similarly to sched_setattr(). The latency priority
is then used to set the offset of the sched_entities of the group.

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Tested-by: K Prateek Nayak <kprateek.nayak@amd.com>
Tested-by: Shrikanth Hegde <sshegde@linux.vnet.ibm.com>

[xMikkkaa: Backported to v4.9. Adapted cgroup handling for latency.nice to v4.9 cpu_files struct, avoiding missing cgroups v2 files.]

Signed-off-by: xMikkkaa <xMikkkaa@users.noreply.github.com>