BACKPORT: sched/fair: Add avg_vruntime
In order to move to an eligibility based scheduling policy it is
needed to have a better approximation of the ideal scheduler.

Specifically, for a virtual time weighted fair queueing based
scheduler the ideal scheduler will be the weighted average of the
individual virtual runtimes (math in the comment).

As such, compute the weighted average to approximate the ideal
scheduler -- note that the approximation is in the individual task
behaviour, which isn't strictly conformant.

Specifically consider adding a task with a vruntime left of center, in
this case the average will move backwards in time -- something the
ideal scheduler would of course never do.

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Tested-by: Shrikanth Hegde <sshegde@linux.vnet.ibm.com>

[xMikkkaa: Backported to v4.9. Manually applied avg_vruntime tracking into older cfs_rq struct and enqueue/dequeue routines.]

Signed-off-by: xMikkkaa <xMikkkaa@users.noreply.github.com>