BACKPORT: sched/fair: Add lag based placement
With the introduction of avg_vruntime, it is possible to approximate
lag (the entire purpose of introducing it in fact). Use this to do lag
based placement over sleep+wake.

Specifically, the FAIR_SLEEPERS thing places things too far to the
left and messes up the deadline aspect of EEVDF.

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Tested-by: Shrikanth Hegde <sshegde@linux.vnet.ibm.com>

[xMikkkaa: Backported to v4.9. Added vlag to sched_entity and integrated update_entity_lag and PLACE_LAG logic into v4.9 codebase.]

Signed-off-by: xMikkkaa <xMikkkaa@users.noreply.github.com>