BACKPORT: sched/eevdf: Sleeper bonus
Add a sleeper bonus hack, but keep it default disabled. This should
allow easy testing if regressions are due to this.
Specifically; this 'restores' performance for things like starve and
stress-futex, stress-nanosleep that rely on sleeper bonus to compete
against an always running parent (the fair 67%/33% split vs the
50%/50% bonus thing).
OTOH this completely destroys latency and hackbench (as in 5x worse).
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Tested-by: Shrikanth Hegde <sshegde@linux.vnet.ibm.com>

xMikkkaa:
This commit introduces an optional sleeper bonus (PLACE_BONUS) for EEVDF,
which is kept default disabled. While it restores performance for certain workloads
that rely on sleeper bonus to compete against an always running parent, it is
notably detrimental to latency in other scenarios.

Signed-off-by: xMikkkaa <xMikkkaa@users.noreply.github.com>