BACKPORT: sched/eevdf: Minimal vavg option
Alternative means of tracking min_vruntime to minimize the deltas
going into avg_vruntime -- note that because vavg move backwards this
is all sorts of tricky.
Also more expensive because of extra divisions... Not found this
convincing.
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Tested-by: Shrikanth Hegde <sshegde@linux.vnet.ibm.com>

xMikkkaa:
This commit provides an alternative means of tracking min_vruntime
(MINIMAL_VA option) to minimize the deltas going into avg_vruntime. It's
noted as an RFC alternative because avg_vruntime moving backwards makes
it tricky and computationally expensive.

Signed-off-by: xMikkkaa <xMikkkaa@users.noreply.github.com>